internal/kubernetes/status/components/cert_manager_csi_driver_spiffe_approver_test.go
package components

import (
	"encoding/json"
	"os"
	"testing"

	"github.com/stretchr/testify/assert"
	"github.com/stretchr/testify/require"
	v1 "k8s.io/api/core/v1"
)

func TestCertManagerCSIDriverSpifferApprover(t *testing.T) {
	var err error
	data, err := os.ReadFile("fixtures/cert-manager-csi-driver-spiffe-approver.json")
	require.NoError(t, err)

	var pod v1.Pod

	err = json.Unmarshal(data, &pod)
	require.NoError(t, err)

	status, err := FindCertManagerCSIDriverSpifferApprover(&pod)
	require.NoError(t, err)
	require.NotNilf(t, status, "expected status to be not nil")

	assert.Equal(t, "cert-manager-csi-driver-spiffe-approver", status.Name())
	assert.Equal(t, "example", status.Namespace())
	assert.Equal(t, "v0.2.0", status.Version())
}
